Added CSS + Table PHP
Table.php will display the current listings of the database whenever the page is loaded/reloaded. This applies to when someone adds/removes someone from the database. Also added some random CSS to give the page some sort of flair instead of a boring white page. The main div portion is a bit fucked up (slightly off-center) at the moment but that can be fixed later.