Updated index/table for separate table
I decided to update the main page for the second iframe table2.php for the additional database table that was added. I also made sure to have the tables print correctly. The only thing now is to get the PHP working in insert.php. I tried to get a test version working but as it turns out, inserting into two different tables is a bitch.